requirements.txt
django==3.0.5
django-widget-tweaks==1.4.8
django-tinymce==2.8.0
cloudinary==1.20.0
#mysqlclient==1.4.6
psycopg2==2.8.5
Pillow==7.1.1
django_cleanup==4.0.0
django-autoslug==1.9.6
xhtml2pdf==0.2.4
gunicorn==20.0.4
dj-database-url==0.5.0
whitenoise==5.0.1
django-heroku==0.3.1